+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-## 1.1.0 (2015-10-14)
+## 1.0.0 (2015-10-14)
 
 - Rewrite all code module
 - Add jwt module for encryption
-- Add middleware method for express for :  auto enrypt/decrypt jsonp request
+- Add middleware method for express for :  auto encode/decode for jsonp request
 - Add middleware method for access checking
+- Add auto encode/decode for json request
+- Add generateAccessToken method